// CaffStore.Frontend/src/app/api/generated/model/registerUserDto.ts
/**
 * Caff Store API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterUserDto { 
    email: string;
    firstName: string;
    lastName: string;
    password: string;
}